+++ Makefile
@@ -24,6 +24,8 @@ LEXER_FILES = $(LEXER_DIR)/tokenizer.c \
 			  $(LEXER_DIR)/handler/expansion_handler.c \
 			  $(LEXER_DIR)/handler/operator_handler.c \
 			  $(LEXER_DIR)/handler/quote_handler.c \
+			  $(LEXER_DIR)/handler/quote_content_handler.c \
+			  $(LEXER_DIR)/handler/quote_expansion.c \
 			  $(LEXER_DIR)/handler/quote_validator.c \
 			  $(LEXER_DIR)/handler/var_expansion_handler.c \
 			  $(LEXER_DIR)/handler/wildcard_handler.c \
