+++ src/lexer/handler/quote_expansion_handler.c
@@ -1,0 +1,56 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        ::::::::            */
+/*   quote_expansion_handler.c                          :+:    :+:            */
+/*                                                     +:+                    */
+/*   By: spyun <spyun@student.codam.nl>               +#+                     */
+/*                                                   +#+                      */
+/*   Created: 2025/02/18 15:24:00 by spyun         #+#    #+#                 */
+/*   Updated: 2025/02/18 15:24:10 by spyun         ########   odam.nl         */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "lexer.h"
+
+static char	*handle_double_quote_expansion(t_tokenizer *tokenizer,
+	char *result)
+{
+	char	*expanded;
+
+	if (ft_strcmp(result, "$?") == 0)
+	{
+		free(result);
+		result = ft_itoa(g_exit_status);
+	}
+	else if (ft_strchr(result, '$'))
+	{
+		expanded = handle_expansion(tokenizer, result);
+		free(result);
+		result = expanded;
+	}
+	return (result);
+}
+
+char	*extract_quoted_content_with_expansion(t_tokenizer *tokenizer,
+	char quote)
+{
+	int		start;
+	char	*result;
+
+	tokenizer->position++;
+	start = tokenizer->position;
+	result = extract_quote_content(tokenizer, quote, start);
+	tokenizer->position++;
+	if (quote == '"')
+		result = handle_double_quote_expansion(tokenizer, result);
+	return (result);
+}
+
+t_quote_state	get_quote_state(t_tokenizer *tokenizer)
+{
+	t_quote_state	state;
+
+	state.in_quote = tokenizer->in_quote;
+	state.quote_char = tokenizer->quote_char;
+	return (state);
+}
